requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
